requirements.txt
# Core NLP Libraries
spacy>=3.7.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
textstat>=0.7.0

# File Processing
PyMuPDF>=1.23.0
python-docx>=0.8.11

# Web Framework
Flask>=2.3.0
Flask-WTF>=1.1.0
Werkzeug>=2.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0.0

# Web Scraping (Optional)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.0.0
flake8>=6.0.0

# Deployment
gunicorn>=21.0.0
python-dotenv>=1.0.0
